fix(frontend): Add show more button to the raw tab
